# Resolve issue 304 regarding Calculator constructor logic